split code of login
